+++ origin/src/Utility/Collection.php
@@ -11,7 +11,7 @@
  * @link        https://www.originphp.com
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Origin\Utils;
+namespace Origin\Utility;
 
 use ArrayIterator;