+++ origin/tests/TestCase/Utility/CollectionTest.php
@@ -11,7 +11,7 @@
  * @link        https://www.originphp.com
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Origin\Test\Utils;
+namespace Origin\Test\Utility;
 
 use Origin\Model\Entity;