+++ origin/tests/TestCase/Utility/NumberTest.php
@@ -12,9 +12,9 @@
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace Origin\Test\Utils;
+namespace Origin\Test\Utility;
 
-use Origin\Utils\Number;
+use Origin\Utility\Number;
 
 class NumberTest extends \PHPUnit\Framework\TestCase
 {