+++ origin/tests/TestCase/Utility/SecurityTest.php
@@ -12,10 +12,10 @@
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace Origin\Test\Utils;
+namespace Origin\Test\Utility;
 
 use Origin\Core\Configure;
-use Origin\Utils\Security;
+use Origin\Utility\Security;
 
 class SecurityTest extends \PHPUnit\Framework\TestCase
 {